refactor out OtherParams
